# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel pour la location d`équipements sportifs",
        "description": "Documentation API Laravel avec Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/equipment": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Récupérer la liste des équipements",
                "description": "Retourner la liste des équipements",
                "operationId": "771e58f407e40ca413f92c85f944e760",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/equipment/{id}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Afficher un équipement",
                "operationId": "546d2e88aec2a0d7af708399e7e20a1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l`équipement",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Équipement non trouvé"
                    }
                }
            }
        },
        "/api/equipment/popularity": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Afficher l`indice de popularité de chaque équipement",
                "description": "Retourne les indices de popularité des équipements",
                "operationId": "59eb1bca862fe1a1b67fdb69c1c335ad",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/rentals": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Afficher le prix moyen des locations de chaque équipement",
                "operationId": "809e242b78331421afe99a338e756cb0",
                "parameters": [
                    {
                        "name": "minDate",
                        "in": "query",
                        "description": "Date minimale requise pour l`affichage",
                        "required": false
                    },
                    {
                        "name": "maxDate",
                        "in": "query",
                        "description": "Date maximale requise pour l`affichage",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reviews/{id}": {
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Supprimer une critique",
                "operationId": "89986343b5dbd2b909c2764aeb7e4096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la critique",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Critique non trouvé"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Récupérer la liste des utilisateurs",
                "description": "Retourner la liste des utilisateurs",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Ajouter un utilisateur",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Samuel"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Lechamplain"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "oui@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "418-321-3210"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur ajouté"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Modifier un utilisateur",
                "operationId": "872a537e8eb13c26c5542c7d73269a80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l`utilisateur",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Samuel"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Lechamplain"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "oui@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "418-321-3210"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Equipment",
            "description": "Equipment"
        },
        {
            "name": "Rentals",
            "description": "Rentals"
        },
        {
            "name": "Reviews",
            "description": "Reviews"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}